Remove the autogenerated starter code. Closes #27
