+++ src/App.vue
@@ -1,83 +1,71 @@
 <template>
-  <v-app>
+  <v-app id="trees-from-seed">
+    <v-navigation-drawer
+      v-model="drawer"
+      app
+    >
+      <v-list v-for="(route, i) in mainRoutes" :key=i nav>
+        <v-list-item :to="route.path" link>
+          <v-list-item-action>
+            <v-icon :color="route.icon.colour">mdi-{{route.icon.name}}</v-icon>
+          </v-list-item-action>
+          <v-list-item-content>
+            <v-list-item-title>{{route.name}}</v-list-item-title>
+          </v-list-item-content>
+        </v-list-item>
+      </v-list>
+    </v-navigation-drawer>
+
     <v-app-bar
       app
       color="primary"
       dark
     >
-      <div class="d-flex align-center">
-        <v-img
-          alt="Vuetify Logo"
-          class="shrink mr-2"
-          contain
-          src="https://cdn.vuetifyjs.com/images/logos/vuetify-logo-dark.png"
-          transition="scale-transition"
-          width="40"
-        />
-
-        <v-img
-          alt="Vuetify Name"
-          class="shrink mt-1 hidden-sm-and-down"
-          contain
-          min-width="100"
-          src="https://cdn.vuetifyjs.com/images/logos/vuetify-name-dark.png"
-          width="100"
-        />
-      </div>
-
-      <v-spacer></v-spacer>
-
-      <v-btn
-        href="https://github.com/vuetifyjs/vuetify/releases/latest"
-        target="_blank"
-        text
-      >
-        <span class="mr-2">Latest Release</span>
-        <v-icon>mdi-open-in-new</v-icon>
-      </v-btn>
+      <v-app-bar-nav-icon @click.stop="drawer = !drawer" />
+      <v-toolbar-title>Trees from Seed</v-toolbar-title>
     </v-app-bar>
 
     <v-content>
-      <div id="nav">
-        |
-        <template v-for="route in mainRoutes">
-          <router-link :to="route.path" :key="route.path">{{route.name}}</router-link> |
-        </template>
-      </div>
-      <router-view/>
+      <v-container
+        class="fill-height"
+        fluid
+      >
+       <router-view></router-view>
+      </v-container>
     </v-content>
+
+    <v-footer
+      app
+      inset
+      color="primary"
+      dark
+    >
+      <v-btn
+        link
+        large
+        outlined
+        href="https://github.com/jimCresswell/trees-from-seed"
+        target="_blank"
+      >
+        <v-icon>mdi-github-circle</v-icon>
+        <span class="ml-2">Source code on GitHub</span>
+      </v-btn>
+    </v-footer>
   </v-app>
 </template>
-
-<style lang="scss">
-#app {
-  text-align: center;
-}
-
-#nav {
-  padding: 30px;
-
-  a {
-    font-weight: bold;
-    color: #2c3e50;
-
-    &.router-link-exact-active {
-      color: #42b983;
-    }
-  }
-}
-</style>
 
 <script>
 import { routes } from './router'
 
 export default {
-  name: 'App',
+  props: {
+  },
 
   // Note: this is data local to the App component.
   // The Vuex store has also been injected in `main.js` and is available to all child components.
   data: () => ({
-    routes
+    routes,
+    drawer: null
   }),
   computed: {
     mainRoutes () {
