+++ src/main.js
@@ -1,6 +1,5 @@
 import Vue from 'vue'
-// import App from './App.vue'
-import App from './NewApp.vue'
+import App from './App.vue'
 import './registerServiceWorker'
 import router from './router'
 import store from './store'
